requirements.txt
streamlit
numpy
pandas
scikit-learn
faiss-cpu
sentence-transformers
PyPDF2
google-genai
